# specification/src/main/resources/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Swagger Microservices Bank
  description: Microservices Bank
  termsOfService: http://swagger.io/terms/
  contact:
    name: Swagger API Team
    email: apiteam@swagger.io
    url: http://swagger.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://editor.swagger.io/api
paths:
...
  /product:
    get:
      tags:
        - PRODUCT
      summary: Returns all product from the system
      description: Returns all product from the system
      operationId: findproduct
      responses:
        '200':
          description: product response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDto'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    post:
      tags:
        - PRODUCT
      summary: Creates a new product in the bank.  Duplicates are not allowed
      description: Creates a new product in the bank.  Duplicates are not allowed
      operationId: addProduct
      requestBody:
        description: ProductDto to add to the bank
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewProductDto'
      responses:
        '200':
          description: product response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  /product/{id}:
    get:
      tags:
        - PRODUCT
      summary: Returns a user based on a identity number
      description: Returns a user based on a identity number
      parameters:
        - name: id
          in: path
          description: Identity number of product to fetch
          required: true
          schema:
            type: string
      responses:
        '200':
          description: product response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    put:
      tags:
        - PRODUCT
      summary: Update product
      description: This can only be done by the logged in product.
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          description: Identity number of product to fetch
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent product
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewProductDto'
          application/xml:
            schema:
              $ref: '#/components/schemas/NewProductDto'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewProductDto'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
        '400':
          description: Invalid identity number
        '404':
          description: Product not found
        '405':
          description: Validation exception
    delete:
      tags:
        - PRODUCT
      summary: Deletes a single product
      description: Deletes a single product
      parameters:
        - name: id
          in: path
          description: id number of product to delete
          required: true
          schema:
            type: string
      responses:
        '204':
          description: product deleted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
...
components:
  schemas:
...
    ProductDto:
      allOf:
        - $ref: '#/components/schemas/NewProductDto'
        - type: object
          required:
            - id
          properties:
            id:
              type: "string"
    NewProductDto:
      allOf:
        - type: object
          required:
            - id
          properties:
            name:
              type: "string"
            category:
              type: "string"
            type:
              type: "string"
            condition:
              $ref: '#/components/schemas/NewConditionDto'
...
    NewConditionDto:
      allOf:
        - type: object
          required:
            - id
          properties:
            hasMaintenanceFee:
              type: "boolean"
            maintenanceFee:
              type: "number"
              format: "double"
            hasMonthlyTransactionLimit:
              type: "boolean"
            monthlyTransactionLimit:
              type: "integer"
              format: "int32"
            hasDailyMonthlyTransactionLimit:
              type: "boolean"
            dailyMonthlyTransactionLimit:
              type: "integer"
              format: "int32"
            productPerPersonLimit:
              type: "integer"
              format: "int32"
            productPerBusinessLimit:
              type: "integer"
              format: "int32"
...
    ErrorDto:
      type: object
      required:
        - message
      properties:
        message:
          type: string